+++ mayan/apps/tags/widgets.py
@@ -1,12 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
 from django import forms
-from django.apps import apps
-from django.template.loader import render_to_string
 from django.utils.html import conditional_escape
-from django.utils.safestring import mark_safe
-
-from .permissions import permission_tag_view
 
 
 class TagFormWidget(forms.SelectMultiple):
@@ -23,29 +18,3 @@ class TagFormWidget(forms.SelectMultiple):
         result['attrs']['data-color'] = self.choices.queryset.get(pk=value).color
 
         return result
-
-
-def widget_document_tags(document, user):
-    """
-    A tag widget that displays the tags for the given document
-    """
-    AccessControlList = apps.get_model(
-        app_label='acls', model_name='AccessControlList'
-    )
-
-    result = ['<div class="tag-container">']
-
-    tags = AccessControlList.objects.filter_by_access(
-        permission_tag_view, user, queryset=document.attached_tags().all()
-    )
-
-    for tag in tags:
-        result.append(widget_single_tag(tag))
-
-    result.append('</div>')
-
-    return mark_safe(''.join(result))
-
-
-def widget_single_tag(tag):
-    return render_to_string('tags/tag_widget.html', {'tag': tag})